Changed connection type from an attribute into an element and made it required.
Moved web service for Reactor connections to "/config/connections".  
Also you can now send it an "id" query parameter to get only connections 
that involve a particular identifier.  It can be either a Reactor or 
connection identifier.

Fixed some deadlock problems involving the new FeedService plug-in

+++ platform/tests/ConfigManagerTests.cpp
@@ -27,5 +27,5 @@ using namespace pion::platform;
 
 BOOST_AUTO_TEST_CASE(checkCreateUUID) {
 	std::string UUID = ConfigManager::createUUID();
-	BOOST_CHECK_EQUAL(UUID.length(), 36);
+	BOOST_CHECK_EQUAL(UUID.length(), static_cast<std::size_t>(36));
 }
